test(multimedia&audioflinger):modify multimedia lib

+++ services/multimedia/speech/process/speech_aec2/include/speech_aec2.h
@@ -23,8 +23,8 @@ typedef struct
     int     shwNlpResdPowAlph;
     int     shwNlpResdPowThd;
     int     shwNlpSmoothGainDod;
-    int     shwNlpBandSortIdx;
-    int     shwNlpBandSortIdxLow;
+    float   shwNlpBandSortIdx;
+    float   shwNlpBandSortIdxLow;
     int     shwNlpTargetSupp;
     int     shwNlpMinOvrd;
     int     shwNlpOvrdHigh;
